Revert previous changes to only keep small improvement of MzTab2
Keep mztab2 format instead of introducing metabolomics.mztab
Do not create metabolomics.py datatype file for the moment, keep everything as is in proteomics.py
Just improve mztab formats by adding descriptions, specific edam_data for mztab2 and save mztab2 files with mztab extension by default
Replace test.mztab2 file
