+++ lib/galaxy/config/sample/datatypes_conf.xml.sample
@@ -391,8 +391,7 @@
     <datatype extension="wiff2.tar" type="galaxy.datatypes.binary:Wiff2Tar" display_in_upload="true"/>
     <datatype extension="mascotxml" type="galaxy.datatypes.proteomics:MascotXML" mimetype="application/xml" display_in_upload="true"/>
     <datatype extension="mztab" type="galaxy.datatypes.proteomics:MzTab" display_in_upload="true"  description="mzTab has been designed to act as a lightweight, tab-delimited file format for mass spec-derived omics data. This is the version 1.x originally designed for proteomics." description_url="https://github.com/HUPO-PSI/mzTab/"/>
-    <!-- mztab2 is just an alias for metabolomics.mztabm for backward-compatibility -->
-    <datatype extension="mztab2" type="galaxy.datatypes.metabolomics:MzTabM" subclass="True" />
+    <datatype extension="mztab2" type="galaxy.datatypes.proteomics:MzTab2" display_in_upload="true" description="mzTab-M is a lightweight, tab-delimited file format for reporting mass spectrometry-based metabolomics results. This is the version 2.x-M of mzTab format which is a refined version of version 1.x to take into account metabolomics aspects." description_url="https://github.com/HUPO-PSI/mzTab/" />
     <datatype extension="mzml" type="galaxy.datatypes.proteomics:MzML" mimetype="application/xml" display_in_upload="true">
         <infer_from suffix="mzml" />
     </datatype>
@@ -461,9 +460,6 @@
     </datatype>
     <datatype extension="colab.tar" auto_compressed_types="gz" type="galaxy.datatypes.binary:CompressedArchive" subclass="true" display_in_upload="true"/>
     <!-- End Proteomics Datatypes -->
-    <!-- Metabolomics Datatypes -->
-    <datatype extension="metabolomics.mztab" type="galaxy.datatypes.metabolomics:MzTabM" display_in_upload="true" description="mzTab-M is a lightweight, tab-delimited file format for reporting mass spectrometry-based metabolomics results. This is the version 2.x-M of mzTab format which is a refined version of version 1.x to take into account metabolomics aspects." description_url="https://github.com/HUPO-PSI/mzTab/" />
-    <!-- End Metabolomics Datatypes -->
     <!-- FlowCytometry -->
     <datatype extension="fcs" type="galaxy.datatypes.flow:FCS" mimetype="application/octet-stream" display_in_upload="true" description="A FCS binary sequence file with a '.fcs' file extension." />
     <datatype extension="flowtext" type="galaxy.datatypes.tabular:Tabular" subclass="true" mimetype="text/tab-separated-values" display_in_upload="true" description="A Flow text file with a '.flowtext' file extension." />
@@ -1333,6 +1329,7 @@
     <sniffer type="galaxy.datatypes.proteomics:MzML"/>
     <sniffer type="galaxy.datatypes.proteomics:MzQuantML"/>
     <sniffer type="galaxy.datatypes.proteomics:MzTab"/>
+    <sniffer type="galaxy.datatypes.proteomics:MzTab2"/>
     <sniffer type="galaxy.datatypes.proteomics:ParamXml"/>
     <sniffer type="galaxy.datatypes.proteomics:MzXML"/>
     <sniffer type="galaxy.datatypes.proteomics:NmrML"/>
@@ -1352,7 +1349,6 @@
     <sniffer type="galaxy.datatypes.proteomics:QCML"/>
     <sniffer type="galaxy.datatypes.proteomics:Wiff"/>
     <sniffer type="galaxy.datatypes.proteomics:PEFF"/>
-    <sniffer type="galaxy.datatypes.metabolomics:MzTabM"/>
     <sniffer type="galaxy.datatypes.molecules:CML"/>
     <sniffer type="galaxy.datatypes.triples:Sbol"/>
     <sniffer type="galaxy.datatypes.xml:GenericXml"/>
