+++ lib/galaxy/datatypes/proteomics.py
@@ -13,6 +13,7 @@ from galaxy.datatypes import data
 from galaxy.datatypes.binary import Binary
 from galaxy.datatypes.data import Text
 from galaxy.datatypes.protocols import (
+    DatasetHasHidProtocol,
     DatasetProtocol,
     HasExtraFilesAndMetadata,
 )
@@ -124,7 +125,7 @@ class MzTab(Text):
     >>> fname = get_test_fname('test.mztab')
     >>> MzTab().sniff(fname)
     True
-    >>> fname = get_test_fname('test.metabolomics.mztab')
+    >>> fname = get_test_fname('test.mztab2')
     >>> MzTab().sniff(fname)
     False
     """
@@ -173,6 +174,55 @@ class MzTab(Text):
             elif columns[0] not in self._sections:
                 return False
         return has_version and found_man_mtd == set(self._man_mtd.keys())
+
+
+class MzTab2(MzTab):
+    """
+    This is version 2.x.x-M of mzTab (mzTab-M) a lightweight, tab-delimited file format for reporting mass spectrometry-based metabolomics results
+    https://github.com/HUPO-PSI/mzTab-M
+
+    >>> from galaxy.datatypes.sniff import get_test_fname
+    >>> fname = get_test_fname('test.mztab2')
+    >>> MzTab2().sniff(fname)
+    True
+    >>> fname = get_test_fname('test.mztab')
+    >>> MzTab2().sniff(fname)
+    False
+    """
+
+    edam_data = "data_4058"
+    file_ext = "mztab2"
+    # section names (except MTD)
+    _sections = ["SMH", "SML", "SFH", "SMF", "SEH", "SME", "COM"]
+    _version_re = r"(2)(\.[0-9]+)?(\.[0-9]+)?-M$"
+    # mandatory metadata fields and list of allowed entries (in lower case)
+    # (or None if everything is allowed)
+    _man_mtd = {"mzTab-ID": None}
+
+    def __init__(self, **kwd):
+        super().__init__(**kwd)
+
+    def display_data(
+        self,
+        trans,
+        dataset: DatasetHasHidProtocol,
+        preview: bool = False,
+        filename: Optional[str] = None,
+        to_ext: Optional[str] = None,
+        **kwd,
+    ):
+        if to_ext == self.file_ext:
+            to_ext = "mztab"
+        return super().display_data(trans, dataset, preview=preview, filename=filename, to_ext=to_ext, **kwd)
+
+    def set_peek(self, dataset: DatasetProtocol, **kwd) -> None:
+        """Set the peek and blurb text"""
+        if not dataset.dataset.purged:
+            dataset.peek = data.get_file_peek(dataset.get_file_name())
+            dataset.blurb = "mzTab-M Format"
+        else:
+            dataset.peek = "file does not exist"
+            dataset.blurb = "file purged from disk"
 
 
 @build_sniff_from_prefix
